+++ src/openMVG/multiview/solver_homography_kernel_test.cpp
@@ -60,7 +60,7 @@ TEST(HomographyKernelTest, Fitting_Unnormalized) {
 
     homography::kernel::UnnormalizedKernel kernel(x, y);
 
-    vector<size_t> samples = {0,1,2,3,4};
+    vector<uint32_t> samples = {0,1,2,3,4};
     for (
       Mat::Index j = 4;
       static_cast<Mat::Index>(samples.size()) < x.cols();
@@ -100,7 +100,7 @@ TEST(HomographyKernelTest, Fitting_Normalized) {
 
     homography::kernel::Kernel kernel(x, y);
 
-    vector<size_t> samples = {0,1,2,3,4};
+    vector<uint32_t> samples = {0,1,2,3,4};
     for (
       Mat::Index j = 4;
       static_cast<Mat::Index>(samples.size()) < x.cols();
