+++ src/openMVG/multiview/solver_resection_p3p.hpp
@@ -368,7 +368,7 @@ class P3P_ResectionKernel_K {
 
   void Fit
   (
-    const std::vector<size_t> &samples,
+    const std::vector<uint32_t> &samples,
     std::vector<Model> *models
   ) const
   {
