+++ src/openMVG/multiview/solver_translation_knownRotation_kernel_test.cpp
@@ -38,7 +38,7 @@ TEST(Translation_knownRotation_Kernel, Multiview) {
 
     openMVG::translation::kernel::TranslationFromKnowRotationKernel kernel(x0, xCam, R_GT);
 
-    const std::vector<size_t> samples = {0,1};
+    const std::vector<uint32_t> samples = {0,1};
     std::vector<Vec3> vec_t;
     kernel.Fit(samples, &vec_t);
