+++ src/openMVG/multiview/two_view_kernel.hpp
@@ -57,8 +57,8 @@ namespace kernel {
 //
 //   1. Kernel::MAX_MODELS
 //   2. Kernel::MINIMUM_SAMPLES
-//   3. Kernel::Fit(std::vector<size_t>, std::vector<Kernel::Model> *)
-//   4. Kernel::Error(size_t, Model) -> error
+//   3. Kernel::Fit(std::vector<uint32_t>, std::vector<Kernel::Model> *)
+//   4. Kernel::Error(uint32_t, Model) -> error
 //
 // The fit routine must not clear existing entries in the vector of models; it
 // should append new solutions to the end.
@@ -78,14 +78,14 @@ class Kernel {
   enum { MAX_MODELS = Solver::MAX_MODELS };
 
   /// Extract required sample and fit model(s) to the sample
-  void Fit(const std::vector<size_t> &samples, std::vector<Model> *models) const {
+  void Fit(const std::vector<uint32_t> &samples, std::vector<Model> *models) const {
     const Mat
       x1 = ExtractColumns(x1_, samples),
       x2 = ExtractColumns(x2_, samples);
     Solver::Solve(x1, x2, models);
   }
   /// Return the error associated to the model and sample^nth point
-  double Error(size_t sample, const Model &model) const {
+  double Error(uint32_t sample, const Model &model) const {
     return ErrorArg::Error(model, x1_.col(sample), x2_.col(sample));
   }
   /// Number of putative point
