+++ src/openMVG/robust_estimation/robust_estimator_lineKernel_test.hpp
@@ -69,7 +69,7 @@ struct LineKernel {
 
   size_t NumSamples() const { return static_cast<size_t> (xs_.cols()); }
 
-  void Fit(const std::vector<size_t> &samples, std::vector<Vec2> *lines) const {
+  void Fit(const std::vector<uint32_t> &samples, std::vector<Vec2> *lines) const {
     assert(samples.size() >= (unsigned int)MINIMUM_SAMPLES);
     // Standard least squares solution.
     const Mat2X sampled_xs = ExtractColumns(xs_, samples);
@@ -77,7 +77,7 @@ struct LineKernel {
     LineSolver::Solve(sampled_xs, lines);
   }
 
-  double Error(size_t sample, const Vec2 &ba) const {
+  double Error(uint32_t sample, const Vec2 &ba) const {
     return pointToLineError::Error(ba, xs_.col(sample));
   }
